+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.3.0
 - Refactor: tonic instead of tikv/grpc-rs
+- Change: Break Change api of auth plugin, support async fn
 - TODO
 
 ### 0.2.6